requirements.txt
openai==1.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
tabula-py==2.8.2
pdfplumber==0.10.3
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
# PDF 표 추출용 패키지
camelot-py[cv]==0.10.1
# RAG 시스템용 패키지
PyMuPDF==1.23.14
chromadb
langchain
langchain-openai
langchain-community
tiktoken
streamlit
